requirements.txt
# Core deep learning framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data manipulation and analysis
numpy>=1.21.0
pandas>=1.3.0

# Machine learning utilities
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Configuration and logging
PyYAML>=6.0
tqdm>=4.62.0

# Image processing
Pillow>=8.3.0

# Optional: For FLOP counting (uncomment if needed)
# ptflops>=0.6.9
# fvcore>=0.1.5

# Jupyter notebook support (for analysis)
jupyter>=1.0.0
ipykernel>=6.0.0

# Additional utilities
pathlib2>=2.3.0  # For Python < 3.4 compatibility
datetime

# For statistical analysis
scipy>=1.7.0

# For better progress bars and logging
rich>=12.0.0

# For reproducibility
random2>=1.0.0
